Fix compiler warnings in GCC9
Things that have been going on for a longer time that I couldn't fix are:
- Warnings for memcopy of things without trivial constructors. These non-trivial constructors could have side effects so the compiler can't guarantee that it's the same as properly constructing these objects. However they don't. And I don't want to hurt performance for that technicality.
- Warnings for assignment constructor over memory that has no trivial destructor. Similar.
- Warning about an unused variable in sti. The variable is used but only depending on preprocessor instructions, so it's best not to remove it in case the conditionals change. I don't really know how to handle this neatly.
- Warnings about unused variable in code generated by Protobuf.

+++ src/utils/polygon.h
@@ -419,7 +419,7 @@ public:
 
     void remove(unsigned int index)
     {
-        POLY_ASSERT(index < size() && index <= std::numeric_limits<int>::max());
+        POLY_ASSERT(index < size() && index <= static_cast<unsigned int>(std::numeric_limits<int>::max()));
         path->erase(path->begin() + index);
     }
 
@@ -610,12 +610,12 @@ public:
 
     PolygonRef operator[] (unsigned int index)
     {
-        POLY_ASSERT(index < size() && index <= std::numeric_limits<int>::max());
+        POLY_ASSERT(index < size() && index <= static_cast<unsigned int>(std::numeric_limits<int>::max()));
         return paths[index];
     }
     ConstPolygonRef operator[] (unsigned int index) const
     {
-        POLY_ASSERT(index < size() && index <= std::numeric_limits<int>::max());
+        POLY_ASSERT(index < size() && index <= static_cast<unsigned int>(std::numeric_limits<int>::max()));
         return paths[index];
     }
     ClipperLib::Paths::iterator begin()
@@ -641,7 +641,7 @@ public:
      */
     void remove(unsigned int index)
     {
-        POLY_ASSERT(index < size() && index <= std::numeric_limits<int>::max());
+        POLY_ASSERT(index < size() && index <= static_cast<unsigned int>(std::numeric_limits<int>::max()));
         if (index < paths.size() - 1)
         {
             paths[index] = std::move(paths.back());
